Add likes functionality to User and Post models, and implement like, update, and delete comment routes

+++ socialmedia/routes/post.py
@@ -1,6 +1,6 @@
-from flask import render_template, url_for, flash, redirect, request
+from flask import render_template, url_for, flash, redirect, request, abort
 from socialmedia import app, db
-from socialmedia.models import Post, Comment
+from socialmedia.models import Post, Comment, likes
 from flask_login import current_user, login_required
 from socialmedia.forms import CommentForm
 
@@ -22,35 +22,52 @@ def get_post_by_id(post_id):
     form = CommentForm()
     post = Post.query.get_or_404(post_id)
     comments = Comment.query.filter_by(post_id=post_id).all()
+    likes = len(post.likes)
     return render_template(
-        "post.html", title="SocialMedia - Post", post=post, comments=comments, form=form
+        "post.html",
+        title="SocialMedia - Post",
+        post=post,
+        comments=comments,
+        form=form,
+        likes=likes,
     )
 
 
-@app.route("/post/<int:post_id>/like", methods=["POST", "GET"])
+@app.route("/post/<int:post_id>/like", methods=["POST"])
 @login_required
 def like_post(post_id):
     post = Post.query.get_or_404(post_id)
-    post.likes += 1
-    db.session.commit()
+    if likes.query.filter_by(user_id=current_user.id, post_id=post.id).first():
+        flash("You already liked this post!", "info")
+    else:
+        like = likes(user_id=current_user.id, post_id=post.id)
+        db.session.add(like)
+        db.session.commit()
+        flash("You liked this post!", "success")
     return redirect(url_for("get_post_by_id", post_id=post_id))
 
 
 @app.route("/post/<int:post_id>/update", methods=["POST"])
 @login_required
 def update_post(post_id):
-    form = request.form["content"]
     post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
+    form = request.form["content"]
+    title = request.form["title"]
     post.content = form
+    post.title = title
     db.session.commit()
     flash("Your post has been updated!", "success")
-    return redirect(url_for("home"))
+    return redirect(url_for("get_post_by_id", post_id=post_id))
 
 
 @app.route("/post/<int:post_id>/delete", methods=["POST"])
 @login_required
 def delete_post(post_id):
     post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
     db.session.delete(post)
     db.session.commit()
     flash("Your post has been deleted!", "success")
